Extract hex map rendering responsibilities
